move cypress tests into different subdirectories and update imports
