+++ tests/cypress/integration/obuk/ConsentAdminTests.ts
@@ -1,13 +1,13 @@
-import { TppIntentPage } from "../pages/tpp/TppIntentPage";
-import { TppLoginPage } from "../pages/tpp/TppLoginPage";
-import { AcpLoginPage } from "../pages/acp/AcpLoginPage";
-import { ConsentPage } from "../pages/consent/ConsentPage";
-import { ErrorPage } from "../pages/ErrorPage";
-import { Credentials } from "../pages/Credentials";
-import { ConsentAdminPage } from "../pages/consent-admin/ConsentAdminPage";
-import { Urls } from "../pages/Urls";
-import { MfaPage } from "../pages/mfa/MfaPage";
-import { EnvironmentVariables } from "../pages/EnvironmentVariables";
+import { TppIntentPage } from "../../pages/tpp/TppIntentPage";
+import { TppLoginPage } from "../../pages/tpp/TppLoginPage";
+import { AcpLoginPage } from "../../pages/acp/AcpLoginPage";
+import { ConsentPage } from "../../pages/consent/ConsentPage";
+import { ErrorPage } from "../../pages/ErrorPage";
+import { Credentials } from "../../pages/Credentials";
+import { ConsentAdminPage } from "../../pages/consent-admin/ConsentAdminPage";
+import { Urls } from "../../pages/Urls";
+import { MfaPage } from "../../pages/mfa/MfaPage";
+import { EnvironmentVariables } from "../../pages/EnvironmentVariables";
 
 describe(`Consent admin app`, () => {
   const tppIntentPage: TppIntentPage = new TppIntentPage();
