+++ tests/cypress/integration/obuk/ConsentSelfServiceTests.ts
@@ -1,18 +1,18 @@
-import {AcpLoginPage} from '../pages/acp/AcpLoginPage';
-import {ConsentPage} from '../pages/consent/ConsentPage';
-import {ErrorPage} from '../pages/ErrorPage';
-import {Credentials} from "../pages/Credentials";
-import {ConsentSelfServicePage} from '../pages/consent-self-service/ConsentSelfServicePage';
-import {Urls} from "../pages/Urls";
-import {MfaPage} from "../pages/mfa/MfaPage";
-import {FinancrooLoginPage} from "../pages/financroo/FinancrooLoginPage";
-import {FinancrooWelcomePage} from "../pages/financroo/FinancrooWelcomePage";
-import {FinancrooAccountsPage} from "../pages/financroo/accounts/FinancrooAccountsPage";
-import {FinancrooInvestmentsPage} from "../pages/financroo/investments/FinancrooInvestmentsPage";
-import {FinancrooContributePage} from "../pages/financroo/investments/FinancrooContributePage";
-import {ConsentSelfServiceApplicationPage} from "../pages/consent-self-service/ConsentSelfServiceApplicationPage";
-import {EnvironmentVariables} from "../pages/EnvironmentVariables"
-import { MockDataRecipientPage } from '../pages/mock-data-recipient/MockDataRecipientPage';
+import {AcpLoginPage} from '../../pages/acp/AcpLoginPage';
+import {ConsentPage} from '../../pages/consent/ConsentPage';
+import {ErrorPage} from '../../pages/ErrorPage';
+import {Credentials} from "../../pages/Credentials";
+import {ConsentSelfServicePage} from '../../pages/consent-self-service/ConsentSelfServicePage';
+import {Urls} from "../../pages/Urls";
+import {MfaPage} from "../../pages/mfa/MfaPage";
+import {FinancrooLoginPage} from "../../pages/financroo/FinancrooLoginPage";
+import {FinancrooWelcomePage} from "../../pages/financroo/FinancrooWelcomePage";
+import {FinancrooAccountsPage} from "../../pages/financroo/accounts/FinancrooAccountsPage";
+import {FinancrooInvestmentsPage} from "../../pages/financroo/investments/FinancrooInvestmentsPage";
+import {FinancrooContributePage} from "../../pages/financroo/investments/FinancrooContributePage";
+import {ConsentSelfServiceApplicationPage} from "../../pages/consent-self-service/ConsentSelfServiceApplicationPage";
+import {EnvironmentVariables} from "../../pages/EnvironmentVariables"
+import { MockDataRecipientPage } from '../../pages/mock-data-recipient/MockDataRecipientPage';
 
 
 describe(`Consent self service app`, () => {
